New: Create API classes for fetch and create/delete/edit
